# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alimentalos API",
        "description": "Official Alimentalos API Documentation",
        "contact": {
            "email": "iantorres@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v.0.6-beta"
    },
    "servers": [
        {
            "url": "https://api.alimentalos.cl",
            "description": "Backend production server"
        }
    ],
    "paths": {
...
        "/users/{user}/groups/{group}/attach": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Attach group of user.",
                "description": "Returns empty array as JSON response.",
                "operationId": "attachUserGroup",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Unique identifier of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Unique identifier of group",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User attached to group successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/pets/{pet}/groups/{group}/attach": {
            "post": {
                "tags": [
                    "Pets"
                ],
                "summary": "Attach group of pet.",
                "description": "Returns empty array as JSON response.",
                "operationId": "attachPetGroup",
                "parameters": [
                    {
                        "name": "pet",
                        "in": "path",
                        "description": "Unique identifier of pet",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Unique identifier of group",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pet attached to group successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/devices/{device}/groups/{group}/attach": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Attach group of device.",
                "description": "Returns empty array as JSON response.",
                "operationId": "attachPetGroup",
                "parameters": [
                    {
                        "name": "device",
                        "in": "path",
                        "description": "Unique identifier of device",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Unique identifier of group",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device attached to group successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/users/{user}/groups/{group}/detach": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Detach group of user.",
                "description": "Returns empty array as JSON response.",
                "operationId": "detachUserGroup",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Unique identifier of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Unique identifier of group",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User detached to group successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/pets/{pet}/groups/{group}/detach": {
            "post": {
                "tags": [
                    "Pets"
                ],
                "summary": "Detach group of pet.",
                "description": "Returns empty array as JSON response.",
                "operationId": "detachPetGroup",
                "parameters": [
                    {
                        "name": "pet",
                        "in": "path",
                        "description": "Unique identifier of pet",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Unique identifier of group",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pet detached to group successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/devices/{device}/groups/{group}/detach": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Detach group of device.",
                "description": "Returns empty array as JSON response.",
                "operationId": "detachPetGroup",
                "parameters": [
                    {
                        "name": "device",
                        "in": "path",
                        "description": "Unique identifier of device",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Unique identifier of group",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device detached to group successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
...
}